Move more code to command.c
