+++ command.c
@@ -22,7 +22,9 @@
 #include <net/net_compat.h>
 #include <net/net_socket.h>
 #endif
+#include <lists/dir_list.h>
 #include <streams/stdin_stream.h>
+#include <streams/file_stream.h>
 #include <string/stdstring.h>
 
 #ifdef HAVE_CONFIG_H
@@ -42,8 +44,10 @@
 #endif
 
 #include "command.h"
+#include "cheat_manager.h"
 #include "content.h"
 #include "dynamic.h"
+#include "list_special.h"
 #include "paths.h"
 #include "verbosity.h"
 #include "version.h"
@@ -938,4 +942,214 @@ bool command_event_save_auto_state(
          "succeeded" : "failed");
 
    return true;
+}
+
+#ifdef HAVE_CHEATS
+void command_event_init_cheats(
+      bool apply_cheats_after_load,
+      const char *path_cheat_db,
+      bsv_movie_t *bsv_movie_state_handle)
+{
+#ifdef HAVE_NETWORKING
+   bool allow_cheats             = !netplay_driver_ctl(
+         RARCH_NETPLAY_CTL_IS_DATA_INITED, NULL);
+#else
+   bool allow_cheats             = true;
+#endif
+#ifdef HAVE_BSV_MOVIE
+   allow_cheats                 &= !(bsv_movie_state_handle != NULL);
+#endif
+
+   if (!allow_cheats)
+      return;
+
+   cheat_manager_alloc_if_empty();
+   cheat_manager_load_game_specific_cheats(path_cheat_db);
+
+   if (apply_cheats_after_load)
+      cheat_manager_apply_cheats();
+}
+#endif
+
+void command_event_load_auto_state(global_t *global)
+{
+   char savestate_name_auto[PATH_MAX_LENGTH];
+   bool ret                        = false;
+#ifdef HAVE_CHEEVOS
+   if (rcheevos_hardcore_active())
+      return;
+#endif
+#ifdef HAVE_NETWORKING
+   if (netplay_driver_ctl(RARCH_NETPLAY_CTL_IS_ENABLED, NULL))
+      return;
+#endif
+
+   savestate_name_auto[0] = '\0';
+
+   fill_pathname_noext(savestate_name_auto, global->name.savestate,
+         ".auto", sizeof(savestate_name_auto));
+
+   if (!path_is_valid(savestate_name_auto))
+      return;
+
+   ret = content_load_state(savestate_name_auto, false, true);
+
+   RARCH_LOG("%s: %s\n%s \"%s\" %s.\n",
+         msg_hash_to_str(MSG_FOUND_AUTO_SAVESTATE_IN),
+         savestate_name_auto,
+         msg_hash_to_str(MSG_AUTOLOADING_SAVESTATE_FROM),
+         savestate_name_auto, ret ? "succeeded" : "failed"
+         );
+}
+
+void command_event_set_savestate_auto_index(
+      settings_t *settings,
+      const global_t *global)
+{
+   size_t i;
+   char state_dir[PATH_MAX_LENGTH];
+   char state_base[PATH_MAX_LENGTH];
+
+   struct string_list *dir_list      = NULL;
+   unsigned max_idx                  = 0;
+   bool savestate_auto_index         = settings->bools.savestate_auto_index;
+   bool show_hidden_files            = settings->bools.show_hidden_files;
+
+   if (!global || !savestate_auto_index)
+      return;
+
+   state_dir[0] = state_base[0]      = '\0';
+
+   /* Find the file in the same directory as global->savestate_name
+    * with the largest numeral suffix.
+    *
+    * E.g. /foo/path/content.state, will try to find
+    * /foo/path/content.state%d, where %d is the largest number available.
+    */
+   fill_pathname_basedir(state_dir, global->name.savestate,
+         sizeof(state_dir));
+
+   dir_list = dir_list_new_special(state_dir, DIR_LIST_PLAIN, NULL,
+         show_hidden_files);
+
+   if (!dir_list)
+      return;
+
+   fill_pathname_base(state_base, global->name.savestate,
+         sizeof(state_base));
+
+   for (i = 0; i < dir_list->size; i++)
+   {
+      unsigned idx;
+      char elem_base[128]             = {0};
+      const char *end                 = NULL;
+      const char *dir_elem            = dir_list->elems[i].data;
+
+      fill_pathname_base(elem_base, dir_elem, sizeof(elem_base));
+
+      if (strstr(elem_base, state_base) != elem_base)
+         continue;
+
+      end = dir_elem + strlen(dir_elem);
+      while ((end > dir_elem) && ISDIGIT((int)end[-1]))
+         end--;
+
+      idx = (unsigned)strtoul(end, NULL, 0);
+      if (idx > max_idx)
+         max_idx = idx;
+   }
+
+   dir_list_free(dir_list);
+
+   configuration_set_int(settings, settings->ints.state_slot, max_idx);
+
+   RARCH_LOG("%s: #%d\n",
+         msg_hash_to_str(MSG_FOUND_LAST_STATE_SLOT),
+         max_idx);
+}
+
+void command_event_set_savestate_garbage_collect(
+      const global_t *global,
+      unsigned max_to_keep,
+      bool show_hidden_files
+      )
+{
+   size_t i, cnt = 0;
+   char state_dir[PATH_MAX_LENGTH];
+   char state_base[PATH_MAX_LENGTH];
+
+   struct string_list *dir_list      = NULL;
+   unsigned min_idx                  = UINT_MAX;
+   const char *oldest_save           = NULL;
+
+   state_dir[0]                      = '\0';
+   state_base[0]                     = '\0';
+
+   /* Similar to command_event_set_savestate_auto_index(),
+    * this will find the lowest numbered save-state */
+   fill_pathname_basedir(state_dir, global->name.savestate,
+         sizeof(state_dir));
+
+   dir_list = dir_list_new_special(state_dir, DIR_LIST_PLAIN, NULL,
+         show_hidden_files);
+
+   if (!dir_list)
+      return;
+
+   fill_pathname_base(state_base, global->name.savestate,
+         sizeof(state_base));
+
+   for (i = 0; i < dir_list->size; i++)
+   {
+      unsigned idx;
+      char elem_base[128];
+      const char *ext                 = NULL;
+      const char *end                 = NULL;
+      const char *dir_elem            = dir_list->elems[i].data;
+
+      elem_base[0]                    = '\0';
+
+      if (string_is_empty(dir_elem))
+         continue;
+
+      fill_pathname_base(elem_base, dir_elem, sizeof(elem_base));
+
+      /* Only consider files with a '.state' extension
+       * > i.e. Ignore '.state.auto', '.state.bak', etc. */
+      ext = path_get_extension(elem_base);
+      if (string_is_empty(ext) ||
+          !string_starts_with_size(ext, "state", STRLEN_CONST("state")))
+         continue;
+
+      /* Check whether this file is associated with
+       * the current content */
+      if (!string_starts_with(elem_base, state_base))
+         continue;
+
+      /* This looks like a valid save */
+      cnt++;
+
+      /* > Get index */
+      end = dir_elem + strlen(dir_elem);
+      while ((end > dir_elem) && ISDIGIT((int)end[-1]))
+         end--;
+
+      idx = string_to_unsigned(end);
+
+      /* > Check if this is the lowest index so far */
+      if (idx < min_idx)
+      {
+         min_idx     = idx;
+         oldest_save = dir_elem;
+      }
+   }
+
+   /* Only delete one save state per save action
+    * > Conservative behaviour, designed to minimise
+    *   the risk of deleting multiple incorrect files
+    *   in case of accident */
+   if (!string_is_empty(oldest_save) && (cnt > max_to_keep))
+      filestream_delete(oldest_save);
+
+   dir_list_free(dir_list);
 }
